Move all modal UI from CanvassAssignmentMap to a separate component
